+++ src/main/java/janggi/controller/Controller.java
@@ -1,6 +1,6 @@
 package janggi.controller;
 
-import janggi.domain.Side;
+import janggi.domain.piece.Side;
 import janggi.domain.board.JanggiBoard;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.generator.DefaultChoPieceGenerator;
