+++ src/main/java/janggi/domain/piece/Elephant.java
@@ -1,6 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.Side;
+import janggi.domain.direction.Vector;
 
 import java.util.List;
