+++ src/main/java/janggi/domain/piece/Guard.java
@@ -1,6 +1,5 @@
 package janggi.domain.piece;
 
-import janggi.domain.Side;
 import java.util.List;
 
 public class Guard extends Piece {
