+++ src/main/java/janggi/domain/piece/Pawn.java
@@ -1,6 +1,5 @@
 package janggi.domain.piece;
 
-import janggi.domain.Side;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
