+++ src/main/java/janggi/service/JanggiBoardService.java
@@ -1,6 +1,6 @@
 package janggi.service;
 
-import janggi.domain.Side;
+import janggi.domain.piece.Side;
 import janggi.domain.board.JanggiBoard;
 import janggi.domain.piece.Piece;
 import janggi.domain.piece.Pieces;
