+++ src/test/java/janggi/domain/board/JanggiBoardTest.java
@@ -1,6 +1,6 @@
 package janggi.domain.board;
 
-import janggi.domain.Side;
+import janggi.domain.piece.Side;
 import janggi.domain.piece.Elephant;
 import janggi.domain.piece.Position;
 import janggi.domain.piece.Rook;
