+++ src/test/java/janggi/domain/piece/PawnTest.java
@@ -1,7 +1,6 @@
 package janggi.domain.piece;
 
 import janggi.domain.ReplaceUnderBar;
-import janggi.domain.Side;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
